// frontend/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Any: { input: any; output: any; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  Duration: { input: any; output: any; }
  Map: { input: any; output: any; }
  Time: { input: any; output: any; }
  UUID: { input: any; output: any; }
  Upload: { input: any; output: any; }
};

export type Action = Node & {
  __typename?: 'Action';
  bazelInvocation: BazelInvocation;
  commandLine?: Maybe<Array<Scalars['String']['output']>>;
  configuration: Configuration;
  endTime?: Maybe<Scalars['Time']['output']>;
  exitCode?: Maybe<Scalars['Int']['output']>;
  failureCode?: Maybe<Scalars['String']['output']>;
  failureMessage?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  label: Scalars['String']['output'];
  startTime?: Maybe<Scalars['Time']['output']>;
  stderr?: Maybe<File>;
  stdout?: Maybe<File>;
  success?: Maybe<Scalars['Boolean']['output']>;
  type?: Maybe<Scalars['String']['output']>;
};
// ...
export type ActionSummary = Node & {
  __typename?: 'ActionSummary';
  actionCacheStatistics?: Maybe<ActionCacheStatistics>;
  actionData?: Maybe<Array<ActionData>>;
  actionsCreated?: Maybe<Scalars['Int']['output']>;
  actionsCreatedNotIncludingAspects?: Maybe<Scalars['Int']['output']>;
  actionsExecuted?: Maybe<Scalars['Int']['output']>;
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  remoteCacheHits?: Maybe<Scalars['Int']['output']>;
  runnerCount?: Maybe<Array<RunnerCount>>;
};
// ...
export type ArtifactMetrics = Node & {
  __typename?: 'ArtifactMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  outputArtifactsFromActionCacheCount?: Maybe<Scalars['Int']['output']>;
  outputArtifactsFromActionCacheSizeInBytes?: Maybe<Scalars['Int']['output']>;
  outputArtifactsSeenCount?: Maybe<Scalars['Int']['output']>;
  outputArtifactsSeenSizeInBytes?: Maybe<Scalars['Int']['output']>;
  sourceArtifactsReadCount?: Maybe<Scalars['Int']['output']>;
  sourceArtifactsReadSizeInBytes?: Maybe<Scalars['Int']['output']>;
  topLevelArtifactsCount?: Maybe<Scalars['Int']['output']>;
  topLevelArtifactsSizeInBytes?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AuthenticatedUser = Node & {
  __typename?: 'AuthenticatedUser';
  bazelInvocations: BazelInvocationConnection;
  displayName?: Maybe<Scalars['String']['output']>;
  externalID: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  userInfo?: Maybe<Scalars['Map']['output']>;
  userUUID: Scalars['UUID']['output'];
};
// ...
export type BazelInvocation = Node & {
  __typename?: 'BazelInvocation';
  actions?: Maybe<Array<Action>>;
  authenticatedUser?: Maybe<AuthenticatedUser>;
  bazelVersion?: Maybe<Scalars['String']['output']>;
  bepCompleted: Scalars['Boolean']['output'];
  build?: Maybe<Build>;
  /** JSON representation of the canonical command line options. */
  canonicalCommandLine?: Maybe<Scalars['Map']['output']>;
  configurations?: Maybe<Array<Configuration>>;
  connectionMetadata?: Maybe<ConnectionMetadata>;
  endedAt?: Maybe<Scalars['Time']['output']>;
  environmentVariables?: Maybe<Scalars['Map']['output']>;
  exitCodeCode?: Maybe<Scalars['Int']['output']>;
  exitCodeName?: Maybe<Scalars['String']['output']>;
  hostname?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  instanceName: InstanceName;
  invocationID: Scalars['UUID']['output'];
  invocationTargets: InvocationTargetConnection;
  metrics?: Maybe<Metrics>;
  numFetches?: Maybe<Scalars['Int']['output']>;
  /** JSON representation of the parsed command line options */
  optionsParsed?: Maybe<Scalars['Map']['output']>;
  /** JSON representation of the original command line options. */
  originalCommandLine?: Maybe<Scalars['Map']['output']>;
  profile?: Maybe<File>;
  sourceControl?: Maybe<Array<SourceControl>>;
  startedAt?: Maybe<Scalars['Time']['output']>;
  tags: InvocationTagConnection;
  username?: Maybe<Scalars['String']['output']>;
};
// ...
export type Build = Node & {
  __typename?: 'Build';
  buildUUID: Scalars['UUID']['output'];
  id: Scalars['ID']['output'];
  instanceName: InstanceName;
  invocations: BazelInvocationConnection;
  tags: BuildTagConnection;
  timestamp: Scalars['Time']['output'];
};
// ...
export type BuildGraphMetrics = Node & {
  __typename?: 'BuildGraphMetrics';
  actionCount?: Maybe<Scalars['Int']['output']>;
  actionCountNotIncludingAspects?: Maybe<Scalars['Int']['output']>;
  actionLookupValueCount?: Maybe<Scalars['Int']['output']>;
  actionLookupValueCountNotIncludingAspects?: Maybe<Scalars['Int']['output']>;
  aspectCounts?: Maybe<Array<BuildGraphAspectCount>>;
  evaluationStats?: Maybe<Array<BuildGraphEvaluationStat>>;
  id: Scalars['ID']['output'];
  inputFileConfiguredTargetCount?: Maybe<Scalars['Int']['output']>;
  metrics?: Maybe<Metrics>;
  otherConfiguredTargetCount?: Maybe<Scalars['Int']['output']>;
  outputArtifactCount?: Maybe<Scalars['Int']['output']>;
  outputFileConfiguredTargetCount?: Maybe<Scalars['Int']['output']>;
  postInvocationSkyframeNodeCount?: Maybe<Scalars['Int']['output']>;
  ruleClassCounts?: Maybe<Array<BuildGraphRuleClassCount>>;
};
// ...
export type Configuration = Node & {
  __typename?: 'Configuration';
  actions?: Maybe<Array<Action>>;
  bazelInvocation: BazelInvocation;
  configurationID: Scalars['String']['output'];
  cpu?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invocationTargets?: Maybe<Array<InvocationTarget>>;
  isTool?: Maybe<Scalars['Boolean']['output']>;
  makeVariables?: Maybe<Scalars['Map']['output']>;
  mnemonic?: Maybe<Scalars['String']['output']>;
  platformName?: Maybe<Scalars['String']['output']>;
};
// ...
export type ConnectionMetadata = Node & {
  __typename?: 'ConnectionMetadata';
  bazelInvocation: BazelInvocation;
  connectionLastOpenAt: Scalars['Time']['output'];
  id: Scalars['ID']['output'];
  timeSinceLastConnectionMillis: Scalars['Int']['output'];
};
// ...
export type CumulativeMetrics = Node & {
  __typename?: 'CumulativeMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  numAnalyses?: Maybe<Scalars['Int']['output']>;
  numBuilds?: Maybe<Scalars['Int']['output']>;
};
// ...
export type DynamicExecutionMetrics = Node & {
  __typename?: 'DynamicExecutionMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  raceStatistics?: Maybe<Array<DynamicExecutionRaceStatistic>>;
};
// ...
export type File = Node & {
  __typename?: 'File';
  actionStderr?: Maybe<Array<Action>>;
  actionStdout?: Maybe<Array<Action>>;
  digest: Digest;
  digestID: Scalars['ID']['output'];
  filePath: FilePath;
  id: Scalars['ID']['output'];
  invocationProfile?: Maybe<Array<BazelInvocation>>;
  testActionOutput?: Maybe<Array<TestResult>>;
};
// ...
export type InstanceName = Node & {
  __typename?: 'InstanceName';
  bazelInvocations?: Maybe<Array<BazelInvocation>>;
  builds?: Maybe<Array<Build>>;
  filePaths?: Maybe<Array<FilePath>>;
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  targets?: Maybe<Array<Target>>;
};
// ...
export type InvocationTagConnection = {
  __typename?: 'InvocationTagConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<InvocationTagEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};
// ...
export type InvocationTargetConnection = {
  __typename?: 'InvocationTargetConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<InvocationTargetEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};
// ...
export type MemoryMetrics = Node & {
  __typename?: 'MemoryMetrics';
  garbageMetrics?: Maybe<Array<GarbageMetrics>>;
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  peakPostGcHeapSize?: Maybe<Scalars['Int']['output']>;
  peakPostGcTenuredSpaceHeapSize?: Maybe<Scalars['Int']['output']>;
  usedHeapSizePostBuild?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Metrics = Node & {
  __typename?: 'Metrics';
  actionSummary?: Maybe<ActionSummary>;
  artifactMetrics?: Maybe<ArtifactMetrics>;
  bazelInvocation?: Maybe<BazelInvocation>;
  buildGraphMetrics?: Maybe<BuildGraphMetrics>;
  cumulativeMetrics?: Maybe<CumulativeMetrics>;
  dynamicExecutionMetrics?: Maybe<DynamicExecutionMetrics>;
  id: Scalars['ID']['output'];
  memoryMetrics?: Maybe<MemoryMetrics>;
  networkMetrics?: Maybe<NetworkMetrics>;
  packageMetrics?: Maybe<PackageMetrics>;
  targetMetrics?: Maybe<TargetMetrics>;
  timingMetrics?: Maybe<TimingMetrics>;
  workerMetrics?: Maybe<Array<WorkerMetrics>>;
  workerPoolMetrics?: Maybe<WorkerPoolMetrics>;
};
// ...
export type NetworkMetrics = Node & {
  __typename?: 'NetworkMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  systemNetworkStats?: Maybe<SystemNetworkStats>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type PackageMetrics = Node & {
  __typename?: 'PackageMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  packageLoadMetrics?: Maybe<Array<PackageLoadMetrics>>;
  packagesLoaded?: Maybe<Scalars['Int']['output']>;
};
// ...
export type RunnerCount = Node & {
  __typename?: 'RunnerCount';
  actionSummary?: Maybe<ActionSummary>;
  actionsExecuted?: Maybe<Scalars['Int']['output']>;
  execKind?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
};
// ...
export type SourceControl = Node & {
  __typename?: 'SourceControl';
  bazelInvocation?: Maybe<BazelInvocation>;
  commit?: Maybe<Scalars['String']['output']>;
  commitURL?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  ref?: Maybe<Scalars['String']['output']>;
  refURL?: Maybe<Scalars['String']['output']>;
  repo?: Maybe<Scalars['String']['output']>;
  repoURL?: Maybe<Scalars['String']['output']>;
};
// ...
export type TimingMetrics = Node & {
  __typename?: 'TimingMetrics';
  actionsExecutionStartInMs?: Maybe<Scalars['Int']['output']>;
  analysisPhaseTimeInMs?: Maybe<Scalars['Int']['output']>;
  cpuTimeInMs?: Maybe<Scalars['Int']['output']>;
  criticalPathTimeInMs?: Maybe<Scalars['Int']['output']>;
  executionPhaseTimeInMs?: Maybe<Scalars['Int']['output']>;
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  wallTimeInMs?: Maybe<Scalars['Int']['output']>;
};
// ...
export type WorkerPoolMetrics = Node & {
  __typename?: 'WorkerPoolMetrics';
  id: Scalars['ID']['output'];
  metrics?: Maybe<Metrics>;
  workerPoolStats?: Maybe<Array<WorkerPoolStats>>;
};
// ...
export type InvocationCompareDataFragment = { __typename?: 'BazelInvocation', id: string, invocationID: any, startedAt?: any | null, endedAt?: any | null, exitCodeName?: string | null, connectionMetadata?: { __typename?: 'ConnectionMetadata', connectionLastOpenAt: any, timeSinceLastConnectionMillis: number } | null, metrics?: { __typename?: 'Metrics', id: string, timingMetrics?: { __typename?: 'TimingMetrics', id: string, wallTimeInMs?: number | null, criticalPathTimeInMs?: number | null, analysisPhaseTimeInMs?: number | null, executionPhaseTimeInMs?: number | null } | null, actionSummary?: { __typename?: 'ActionSummary', id: string, runnerCount?: Array<{ __typename?: 'RunnerCount', id: string, name?: string | null, actionsExecuted?: number | null }> | null } | null } | null, numTotal: { __typename?: 'InvocationTargetConnection', totalCount: number }, numSuccessful: { __typename?: 'InvocationTargetConnection', totalCount: number }, numSkipped: { __typename?: 'InvocationTargetConnection', totalCount: number } } & { ' $fragmentName'?: 'InvocationCompareDataFragment' };